+++ config/autoload.php
@@ -24,15 +24,15 @@ ClassLoader::addNamespaces(array
 ClassLoader::addClasses(array
 (
 	// Classes
-	'HeimrichHannot\Ajax\Ajax'                        => 'system/modules/ajax/classes/Ajax.php',
-	'HeimrichHannot\Ajax\Hooks'                       => 'system/modules/ajax/classes/Hooks.php',
-	'HeimrichHannot\Ajax\AjaxAction'                  => 'system/modules/ajax/classes/AjaxAction.php',
-	'HeimrichHannot\Ajax\AjaxToken'                   => 'system/modules/ajax/classes/AjaxToken.php',
-	'HeimrichHannot\Ajax\Exception\AjaxExitException' => 'system/modules/ajax/classes/Exception/AjaxExitException.php',
-	'HeimrichHannot\Ajax\Response\ResponseRedirect'   => 'system/modules/ajax/classes/Response/ResponseRedirect.php',
-	'HeimrichHannot\Ajax\Response\Response'           => 'system/modules/ajax/classes/Response/Response.php',
-	'HeimrichHannot\Ajax\Response\Response404'        => 'system/modules/ajax/classes/Response/Response404.php',
-	'HeimrichHannot\Ajax\Response\ResponseData'       => 'system/modules/ajax/classes/Response/ResponseData.php',
-	'HeimrichHannot\Ajax\Response\ResponseError'      => 'system/modules/ajax/classes/Response/ResponseError.php',
-	'HeimrichHannot\Ajax\Response\ResponseSuccess'    => 'system/modules/ajax/classes/Response/ResponseSuccess.php',
+	'HeimrichHannot\Ajax\Ajax'                        => 'system/modules/ajax/src/Ajax.php',
+	'HeimrichHannot\Ajax\Hooks'                       => 'system/modules/ajax/src/Hooks.php',
+	'HeimrichHannot\Ajax\AjaxAction'                  => 'system/modules/ajax/src/AjaxAction.php',
+	'HeimrichHannot\Ajax\AjaxToken'                   => 'system/modules/ajax/src/AjaxToken.php',
+	'HeimrichHannot\Ajax\Exception\AjaxExitException' => 'system/modules/ajax/src/Exception/AjaxExitException.php',
+	'HeimrichHannot\Ajax\Response\ResponseRedirect'   => 'system/modules/ajax/src/Response/ResponseRedirect.php',
+	'HeimrichHannot\Ajax\Response\Response'           => 'system/modules/ajax/src/Response/Response.php',
+	'HeimrichHannot\Ajax\Response\Response404'        => 'system/modules/ajax/src/Response/Response404.php',
+	'HeimrichHannot\Ajax\Response\ResponseData'       => 'system/modules/ajax/src/Response/ResponseData.php',
+	'HeimrichHannot\Ajax\Response\ResponseError'      => 'system/modules/ajax/src/Response/ResponseError.php',
+	'HeimrichHannot\Ajax\Response\ResponseSuccess'    => 'system/modules/ajax/src/Response/ResponseSuccess.php',
 ));